refactored receipt parsing
